target position wip, type long updated in struct index and target pos

+++ libft/includes/libft.h
@@ -6,7 +6,7 @@
 /*   By: rcochran <rcochran@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/13 15:28:08 by rcochran          #+#    #+#             */
-/*   Updated: 2025/02/18 12:40:04 by rcochran         ###   ########.fr       */
+/*   Updated: 2025/02/19 15:55:48 by rcochran         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -98,5 +98,5 @@ char				*ft_convert_base(char *nbr, char *base_from, char *base_to);
 int					is_in_base(char c, char *base);
 int					is_base_valid(char *base);
 
-void				ft_swap(int *a, int *b);
+void				ft_swap(long *a, long *b);
 #endif
